[S12P21E206-187] 정산하기 기능 오더,셀러 구현

+++ back/funding/src/main/java/com/ssafy/funding/mapper/FundingMapper.java
@@ -59,5 +59,5 @@ public interface FundingMapper {
     List<Funding> findByStatusAndEventSent(@Param("eventSent") Boolean eventSent);
 
     // settlement_completed 플래그 업데이트
-    int updateSettlementCompleted(@Param("fundingId") Long fundingId, @Param("eventSent") Boolean eventSent);
+    int updateSettlementCompleted(@Param("fundingId") int fundingId, @Param("eventSent") Boolean eventSent);
 }
